se quita el campo amount de la vista de pagos, se usara solo el precio del plan

+++ domain/src/main/java/com/example/cleanarchitecture/domain/domain/entity/Plan.kt
@@ -11,7 +11,7 @@ data class Plan(
 
     val id: String? =null,
     val name:String="",
-    val price: Float = 0f ,
+    val price: Double = 0.0 ,
     val downloadSpeed: String="",
     val uploadSpeed: String="",
